fix: apply even/odd classes correctly when rows are grouped
BREAKING CHANGE: Previously `CellContext.rowIndex` was a `string` if the row was inside a group.
Now `CellContext.rowIndex` is always a number.
Either containing the index of the row or if the row is inside a group, the index of the group.
To access the index value of a row within a group, use the new`CellContext.rowInGroupIndex`.