fix: only run disabled check on rows
We previously ran `disableRowCheck` on groups instead of each row
if rows were grouped.
With this change, we only run the check on rows.
Besides this, the uncommon `disableRow$` subject was removed from template
contexts and replaced with a proper `disabled` field.
This simplifies the usage and prevents wrong expectations.
Although even shown in our example, calling `next` on the subject works
but the value will be overridden on the next CD cycle, so it was pointless.
Replacing the subject with a normal boolean resolves this problem.

Being able to subscribe to the subject is also not needed.
An application always knows whether a row was disabled
because they set it.

BREAKING CHANGE: The row detail context (`RowDetailContext`) no longer contains `disableRow$?: Observable<boolean>`.
Use the new `disabled: boolean` value instead.

BREAKING CHANGE: The cell context (`CellContext`) no longer contains `disableRow$?: Observable<boolean>`.
Use the new `disabled: boolean` value instead.

BREAKING CHANGE: `disableRowCheck` will only be called with actual rows.
Although documented otherwise, in the case of groups, the `disableRowCheck`
was only called for groups instead of each row inside that group.
This allows disabling single rows inside a group and not only entire groups.

To update the disabled state of a row just update the row itself instead of using the previous `disableRow$` subject.