fix: show the correct number of pages when using row grouping
When using grouping and `limit` (maximum number of rows per page) together,
the table will count the groups, not the rows.
So a limit 4 means that 4 groups are rendered per page.
Yet, the pager was counting the number rows within a group.
With this fix, the pager will be aligned with the body,
and also count the groups.