fix: support parallel usage of grouping and details
Previously, row-grouping and details shared the same expansion state.
So it was impossible to use them in parallel.
This commit separates them into two different fields.
It is done by duplicating the relevant fields in the body-component,
and by splitting row-wrapper into an actual row-wrapper for managing the details
and a group-wrapper, for managing groups.
This is a preparation to split the group renderîng from the body component.

Closes #147