fix: always call `<ngx-datatable-row-detail [rowHeight]="function">` with correct index
Previously, when using `<ngx-datatable-row-detail [rowHeight]="function">`, the function
was called sporadically with a wrong index.
This is due to a bug in the body-component, where the rendered offset-index was passed to the
cache. But the cache is operating on all rows, not the rendered ones.
So instead, we need to pass only the offset, if external paging is enabled.