refactor: convert `DataTableColumnDirective` to signal inputs
As a main change, the `DataTableColumnDirective` is now using signal inputs.
In order to get this work, instead of implementing `TableColumn`,
the directive now has a property called `column`, which contains combined values.
The positive side effect is that subscribing to changes is now handled by signals,
and we no longer need the related service. We just use an effect.